backend-django/requirements.txt
Django==5.1.1
djangorestframework==3.15.2
django-cors-headers==4.4.0
requests==2.32.3
reportlab==4.2.5


